pyproject.toml
[project]
name = "ai-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=38.2.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.34",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.8",
    "langgraph-cli[inmem]>=0.4.2",
    "langsmith>=0.4.31",
    "mcp>=1.15.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "strawberry-graphql>=0.287.0",
]
